Add new and extended.

+++ install/sql/table_points_comp_report.sql
@@ -5,8 +5,8 @@ CREATE TABLE /*_*/points_comp_report (
   `points` int(8) NOT NULL,
   `month_start` int(14) NOT NULL DEFAULT '0',
   `month_end` int(14) NOT NULL DEFAULT '0',
-  `new` tinyint(1) NOT NULL DEFAULT '0',
-  `extended` tinyint(1) NOT NULL DEFAULT '0',
+  `new` int(10) NOT NULL DEFAULT '0',
+  `extended` int(10) NOT NULL DEFAULT '0',
   `comp_expires` int(14) NOT NULL DEFAULT '0'
 ) /*$wgDBTableOptions*/;
